Don't take EPOCHREALTIME availability as proof of the shell being bash
Presently, there are two working implementations of the _update_time()
function, one of which is intended for bash while the other is intended
for sh. The bash implementation is chosen in the case that the
'EPOCHREALTIME' variable is found to be functional.

However, it is possible for busybox to be built with support for both
the 'EPOCHSECONDS' and 'EPOCHREALTIME' variables. Therefore, it cannot
be assumed that the shell is bash on that basis alone.

Address this issue by also testing whether the 'BASH' variable is
non-empty in the course of selecting the implementation.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>